Remove outdated run-tests files, update travis file.
Python 3 compatability pass.

Switch from endswith to contains.

Fix tests.
