--- run-tests-2.4
@@ -1,1 +1,0 @@
-clear && find ./ -name ".coverage" -delete && find ./ -name "*.pyc" -delete && nosetests-2.4 -vd --stop --with-cover --cover-package=shotgun_api3
