+++ shotgun_api3/lib/httplib2/python2/cacerts.txt
