+++ tests/base.py
@@ -1,15 +1,17 @@
 """Base class for Shotgun API tests."""
+from __future__ import absolute_import
 import os
 import re
 import unittest
-from ConfigParser import ConfigParser
+from six.moves.configparser import ConfigParser
 
 
-import mock
+from . import mock
 
 import shotgun_api3 as api
 from shotgun_api3.shotgun import json
 from shotgun_api3.shotgun import ServerCapabilities
+import six
 
 CONFIG_PATH = 'tests/config'
 
@@ -133,8 +135,18 @@ class MockTestBase(TestBase):
         if not isinstance(self.sg._http_request, mock.Mock):
             return
 
-        if not isinstance(data, basestring):
-            data = json.dumps(data, ensure_ascii=False, encoding="utf-8")
+        if not isinstance(data, six.string_types):
+            if six.PY2:
+                data = json.dumps(
+                    data,
+                    ensure_ascii=False,
+                    encoding="utf-8"
+                )
+            else:
+                data = json.dumps(
+                    data,
+                    ensure_ascii=False,
+                )
 
         resp_headers = { 'cache-control': 'no-cache',
                          'connection': 'close',
@@ -157,7 +169,7 @@ class MockTestBase(TestBase):
         """Asserts _http_request is called with the method and params."""
         args, _ = self.sg._http_request.call_args
         arg_body = args[2]
-        assert isinstance(arg_body, basestring)
+        assert isinstance(arg_body, six.string_types)
         arg_body = json.loads(arg_body)
 
         arg_params = arg_body.get("params")
@@ -359,7 +371,7 @@ def _find_or_create_entity(sg, entity_type, data, identifyiers=None):
     @returns dicitonary of the entity values
     '''
     identifyiers = identifyiers or ['name']
-    fields = data.keys()
+    fields = list(data.keys())
     filters = [[key, 'is', data[key]] for key in identifyiers]
     entity = sg.find_one(entity_type, filters, fields=fields)
     entity = entity or sg.create(entity_type, data, return_fields=fields)
