+++ tests/tests_proxy.py
@@ -1,6 +1,7 @@
 #! /opt/local/bin/python
+from __future__ import absolute_import
 import sys
-import base
+from . import base
 import shotgun_api3 as api
